requirements.txt
numpy
scikit-image
matplotlib
pillow